se refactoriza el codigo y se agregan funcionalidades
